+++ src/web/Routes.jsx
@@ -72,8 +72,8 @@ import ReportDetailsPage from 'web/pages/reports/DetailsPage';
 import ReportsPage from 'web/pages/reports/ReportListPage';
 import ResultDetailsPage from 'web/pages/results/DetailsPage';
 import ResultsPage from 'web/pages/results/ListPage';
-import RoleDetailsPage from 'web/pages/roles/DetailsPage';
-import RolesPage from 'web/pages/roles/ListPage';
+import RoleDetailsPage from 'web/pages/roles/RoleDetailsPage';
+import RolesPage from 'web/pages/roles/RoleListPage';
 import ScanConfigDetailsPage from 'web/pages/scanconfigs/DetailsPage';
 import ScanConfigsPage from 'web/pages/scanconfigs/ListPage';
 import ScannerDetailsPage from 'web/pages/scanners/DetailsPage';
